Update README.mediawiki
Most of these proposed edits are syntax changes when the length or construction of a sentence made the point confusing.... also a few grammar, spelling and spacing corrections.  In one case a sentence was duplicated, appearing at the top of the abstract and also at the bottom of the Introduction.