Update README.mediawiki
I reworded the sentence: Sovereign's codebase is able to deliver a mobile and desktop application for voters and organizations willing to implement a blockchain based democracy.  As written, it reads like a shaky use case proposition to me, to say that DEF's software is for voters and orgs 'willing to implement'.....it may just be a syntax issue, however this struck me as a sort of keystone sentence, saying who the sw is for, & therefore important to get right.