Editing for readability
Line 30 - the syntax is awkward in this sentence. 

Current: 
Our aim is to help echo a mission encompassed in all of humanity as we see the need to make of our shared home an utopia (not doing so could otherwise be rather expensive if not fatal). 

Rewrite: 
Our mission echoes across generations and continents, it is a mission for all of humanity, to heed the call to make of our shared home a utopia….or lose the hope of home and humanity altogether.