edited for syntax
Line 73 Current
We can consider elections implemented by states, provinces and city municipalities as democracies where we are reduced to being passive recipients of a monologue. Citizens are called in-between substantially long periods of time, during elections, to provide a basic input: essentially accept or reject players in the same system. This is the bandwidth of the legacy system that is our so-called modern democracies.

Rewrite: 
We can consider elections implemented by states, provinces and city municipalities as democracies in which citizens are reduced to mere passive recipients of monologues and allowed only the most intermittent and basic participation, i.e. called upon every 4-6 years to essentially accept or reject players in the same system --- the legacy system of our so-called modern democracies.