148: typo fix
was
Blockchains operating in cyberspace are giving rise to a fourth kind: the networked individual. It’s not a far fetched possibility: conquering personal sovereignty is already a reality for those who run their finances with bitcoin and other crypto hodlings. 

new
Blockchains operating in cyberspace are giving rise to a fourth kind: the networked individual. It’s not a far fetched possibility: conquering personal sovereignty is already a reality for those who run their finances with bitcoin and other crypto holdings.