added reference connecting democracy to UBI
was: The ability to develop a reliable self-sovereign identity validation process not only guarantees the legitimate value for ''votes'' to express social choice but also establishes the bedrock infrastructure required to make a ''Universal Basic Income'' (UBI) mechanism that can reach everyone on Earth. Therefore in order to consolidate the political and financial logic able to establish a borderless democracy, once a ''Proof of Identity'' is validated by peers the distribution mechanism triggered for ''votes'' will be based on time as a ''UBI''.

now: The ability to develop a reliable self-sovereign identity validation process not only guarantees the legitimate value for ''votes'' to express social choice but also establishes the bedrock infrastructure required to make a ''Universal Basic Income'' (UBI) mechanism that can reach everyone on Earth. The symbiotic relationship of UBI and democracy has been well substantiated. According to research presented at Basic Income Earth Network (Munich September 2012), the implementation of basic income [http://basicincome.org/bien/pdf/munich2012/Choi.pdf can greatly contribute to realizing the principles of democracy as well as the establishment of its substantial foundation]. Therefore in order to consolidate the political and financial logic able to establish a borderless democracy, once a ''Proof of Identity'' is validated by peers the distribution mechanism triggered for ''votes'' will be based on time as a ''UBI''.