from anyone to no one...a big little correction
somehow it seems this sentence is saying the opposite of what it intends - it intends to say there is no accountability to anyone but the board.    But what it says is that they are accountable to everyone except the board!  

line 108 was
It is in their interest to gather as much information as possible to profile people in order to stay competitive in the attention market and both companies filter information fed to users with algorithms accountable to anyone but their own board. 

changed to

It is in their interest to gather as much information as possible to profile people in order to stay competitive in the attention market and both companies filter information fed to users with algorithms accountable to no one but their own board.